Extract CovidStatsPanel
