+++ src/components/CovidStats/CovidStatsFigure.vue
